transition_system fix
